+++ sdk/dotnet/S3/Inputs/BucketServerSideEncryptionConfigurationRuleGetArgs.cs
@@ -18,6 +18,12 @@ namespace Pulumi.Aws.S3.Inputs
         [Input("applyServerSideEncryptionByDefault", required: true)]
         public Input<Inputs.BucketServerSideEncryptionConfigurationRuleApplyServerSideEncryptionByDefaultGetArgs> ApplyServerSideEncryptionByDefault { get; set; } = null!;
 
+        /// <summary>
+        /// Whether or not to use [Amazon S3 Bucket Keys](https://docs.aws.amazon.com/AmazonS3/latest/dev/bucket-key.html) for SSE-KMS.
+        /// </summary>
+        [Input("bucketKeyEnabled")]
+        public Input<bool>? BucketKeyEnabled { get; set; }
+
         public BucketServerSideEncryptionConfigurationRuleGetArgs()
         {
         }